requirements.txt
pyserial>=3.0,<4
modbus_tk>=0.4.3
redis>=2.10.5,<3
requests>=2.9,<3
prometheus-client==0.0.21
